编写 Order 和 Gift DTO 层

+++ model/order.go
@@ -5,7 +5,7 @@ import "time"
 type Order struct {
 	ID        int64      `gorm:"primaryKey"`        // 订单 ID
 	UserID    int64      `gorm:"column:user_id"`    // 用户 ID
-	GiftID    int        `gorm:"column:gift_id"`    // 礼物 ID
+	GiftID    int64      `gorm:"column:gift_id"`    // 礼物 ID
 	Count     int        `gorm:"column:count"`      // 购买数量
 	CreatedAt time.Time  `gorm:"column:created_at"` // 创建时间
 	UpdatedAt time.Time  `gorm:"column:updated_at"` // 更新时间
